Build the base script
This script will remove the com.apple.quarantine extended attribute
from an application for a user that does not have admin rights. The
idea is to have this run from JAMF's Self Service. This is for people
that are placing Applications in their home dirs to run but are not
signed by the developer.